Update domain models for consistency with Core
Specifically, this updates the Client template to
address the NPE issue with the passwordEncoder
as seen in GPSPRINGSECURITYCORE-250.

This also fixes an issue where the unit tests were
polluting the security config resulting in the
random failures caused by the plugin being inactive.

+++ grails-app/domain/test/oauth2/UserRole.groovy
@@ -15,7 +15,7 @@ class UserRole implements Serializable {
 		}
 
 		other.user?.id == user?.id &&
-			other.role?.id == role?.id
+		other.role?.id == role?.id
 	}
 
 	int hashCode() {
@@ -32,30 +32,63 @@ class UserRole implements Serializable {
 		}.get()
 	}
 
+	static boolean exists(long userId, long roleId) {
+		UserRole.where {
+			user == User.load(userId) &&
+			role == Role.load(roleId)
+		}.count() > 0
+	}
+
 	static UserRole create(User user, Role role, boolean flush = false) {
-		new UserRole(user: user, role: role).save(flush: flush, insert: true)
+		def instance = new UserRole(user: user, role: role)
+		instance.save(flush: flush, insert: true)
+		instance
 	}
 
 	static boolean remove(User u, Role r, boolean flush = false) {
+		if (u == null || r == null) return false
 
 		int rowCount = UserRole.where {
 			user == User.load(u.id) &&
 			role == Role.load(r.id)
 		}.deleteAll()
 
+		if (flush) { UserRole.withSession { it.flush() } }
+
 		rowCount > 0
 	}
 
-	static void removeAll(User u) {
+	static void removeAll(User u, boolean flush = false) {
+		if (u == null) return
+
 		UserRole.where {
 			user == User.load(u.id)
 		}.deleteAll()
+
+		if (flush) { UserRole.withSession { it.flush() } }
 	}
 
-	static void removeAll(Role r) {
+	static void removeAll(Role r, boolean flush = false) {
+		if (r == null) return
+
 		UserRole.where {
 			role == Role.load(r.id)
 		}.deleteAll()
+
+		if (flush) { UserRole.withSession { it.flush() } }
+	}
+
+	static constraints = {
+		role validator: { Role r, UserRole ur ->
+			if (ur.user == null) return
+			boolean existing = false
+			UserRole.withNewSession {
+				existing = UserRole.exists(ur.user.id, r.id)
+			}
+			if (existing) {
+				return 'userRole.exists'
+			}
+		}
 	}
 
 	static mapping = {
